Make PatchExternalModules cacheable again
There where two things going on:

First, we had the version-info on the coreRuntimeClasspath
which is an input to patchExternalModules.

Secondly, we had externalModulesRuntime as `NAME_ONLY`
inputs to the PatchExternalModules task. This configuration contains
the core configuration which now has all the jars with the timestamp
in it - so it was never up to date with respect to `NAME_ONLY`.

+review REVIEW-6331